% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frdistogives.R
\name{basicfreqdist}
\alias{basicfreqdist}
\title{Automated Cumulative Frequency Plots for Grouped Distribution}
\usage{
basicfreqdist(lower_Class_limit, upper_Class_limit, freq)
}
\arguments{
\item{lower_Class_limit, upper_Class_limit, freq}{numeric vector}
}
\value{
frtable data.frame CUMULATIVE FREQUENCY DISTRIBUTION TABLE
}
\description{
Creates a FREQUENCY DISTRIBUTION TABLE and gives a cumulative frequency plot for UNIFORMLY DISTRIBUTED FREQUENCY DISTRIBUTION wherein the data is in the form of 3 arguments lower class limits upper class limits and frequencies
}
\examples{
basicfreqdist(seq(0,50,by=10),seq(10,60,by=10),c(34,53,44,23,39,40))
}
